Week_2_BTTL/requirements.txt
flask
PyJWT

